+++ cdot65/prisma_access/plugins/module_utils/api_spec.py
@@ -68,7 +68,7 @@ class PrismaAccessSpec:
         )
 
     @staticmethod
-    def addresses_spec():
+    def address_spec():
         """Return the address object spec."""
         return dict(
             description=dict(